Chat en api

+++ SwapifyAPI/src/main/kotlin/com/swapify/swapifyapi/services/ChatServiceInterface.kt
@@ -4,6 +4,6 @@ import com.swapify.swapifyapi.model.dto.ChatMessageDTO
 import com.swapify.swapifyapi.model.entities.Message
 
 interface ChatServiceInterface {
-    fun saveMessage(roomId: Long, dto: ChatMessageDTO): Message
-    fun findMessages(roomId: Long): List<Message>
+    fun saveMessage(roomId: Int, dto: ChatMessageDTO): Message
+    fun findMessages(roomId: Int): List<Message>
 }
